requirements.txt
anyio==3.6.1
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
ecdsa==0.18.0
fastapi==0.79.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
PyJWT==2.4.0
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette>=0.25.0
typing_extensions==4.3.0
uvicorn==0.18.2
